ecomm/requirements.txt
django==3.0.3
djangorestframework==3.11.0
django-rest-auth==0.9.5
django-allauth==0.41.0
psycopg2==2.8.4
djangorestframework-jwt==1.11.0
